pyproject.toml
[project]
name = "flame"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.2",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
